split components into different files
