+++ src/modules/Music.py
@@ -1,29 +1,13 @@
-#!/usr/bin/python3
-
 import asyncio
-
-import youtube_dl
-import os
-from dotenv import load_dotenv
-import requests
 import time
+
 import discord
+import requests
+import youtube_dl
 from discord.ext import commands
-
-# Load Token
-load_dotenv()
-TOKEN=os.getenv('BOT_TOKEN')
-UID1=os.getenv('UID1')
-UID2=os.getenv('UID2')
-UID3=os.getenv('UID3')
-UID4=os.getenv('UID4')
-
-# Auth UIDs
-AUTHENTICATED={int(UID1), int(UID2), int(UID3), int(UID4)}
 
 # Suppress noise about console usage from errors
 youtube_dl.utils.bug_reports_message = lambda: ""
-
 
 ytdl_format_options = {
     "format": "bestaudio/best",
@@ -146,19 +130,3 @@ class Music(commands.Cog):
                 raise commands.CommandError("Author not connected to a voice channel.")
         elif ctx.voice_client.is_playing():
             ctx.voice_client.stop()
-
-
-bot = commands.Bot(
-    command_prefix=commands.when_mentioned_or("!"),
-    description="Relatively simple music bot example",
-)
-
-
-@bot.event
-async def on_ready():
-    print(f"Logged in as {bot.user} (ID: {bot.user.id})")
-    print("------")
-
-
-bot.add_cog(Music(bot))
-bot.run(TOKEN)
